+++ src/main/java/faang/school/accountservice/service/BalanceService.java
@@ -1,10 +1,7 @@
 package faang.school.accountservice.service;
 
 import faang.school.accountservice.dto.BalanceDto;
-import faang.school.accountservice.dto.PendingDto;
 import faang.school.accountservice.entity.Balance;
-
-import java.util.List;
 
 public interface BalanceService {
 
@@ -12,13 +9,9 @@ public interface BalanceService {
 
     void update(BalanceDto balanceDto);
 
-    void paymentAuthorization(PendingDto pendingDto);
-
-    void clearPayment(List<PendingDto> pendingDto);
-
-    void clearPayment(PendingDto pendingDto);
-
     BalanceDto getBalance(long accountId);
 
-    BalanceDto getBalanceByAccountId(long accountId);
+    BalanceDto getBalanceDtoByAccountId(long accountId);
+
+    Balance getBalanceByAccountId(long accountId);
 }
