+++ src/main/java/faang/school/accountservice/service/PaymentService.java
@@ -1,0 +1,10 @@
+package faang.school.accountservice.service;
+
+import faang.school.accountservice.dto.PendingDto;
+
+public interface PaymentService {
+
+    void paymentAuthorization(PendingDto pendingDto);
+
+    void clearPayment(PendingDto pendingDto);
+}
